@@ -1,14 +1,29 @@
-require File.expand_path(File.dirname(__FILE__) + '/../spec_helper')
+require 'spec_helper'
 
-describe FaqsController do
+describe Spree::FaqsController do
 
-  it "should response to index" do
-    QuestionCategory.should_receive(:all).and_return(:categories)
+  let!(:question_category) { create(:question_category) }
+  let!(:question) { create(:question, question_category: question_category) }
 
-    get :index
+  before { controller.stub spree_current_user: nil }
 
-    assigns(:categories).should eql(:categories)
-    response.should be_success
+  context '#index' do
+    specify do
+      categories = [
+        question_category,
+        create(:question_category),
+        create(:question_category)
+      ]
+
+      spree_get :index
+      expect(assigns(:categories)).to eq categories
+      expect(response).to be_success
+    end
   end
 
+  context '#default_title' do
+    it 'return default title' do
+      expect(subject.default_title).to eq Spree.t(:frequently_asked_questions, scope: :spree_faq)
+    end
+  end
 end
