@@ -1,7 +1,6 @@
-Rails.application.routes.draw do
-  match :faq, :to => 'faqs#index', :as => 'faq'
-
-  namespace :admin do 
+Spree::Core::Engine.add_routes do
+  namespace :admin do
     resources :question_categories
   end
+  get 'faq', to: 'faqs#index'
 end
