@@ -1,3 +1,4 @@
 require 'spree_core'
-require 'spree_faq_hooks'
+require 'spree_faq/version'
 require 'spree_faq/engine'
+require 'coffee_script'
