@@ -1,28 +1,25 @@
-# This file is copied to ~/spec when you run 'ruby script/generate rspec'
-# from the project root directory.
-ENV["RAILS_ENV"] ||= 'test'
-require File.expand_path("../test_app/config/environment", __FILE__)
-require 'rspec/rails'
-
-# Requires supporting files with custom matchers and macros, etc,
-# in ./support/ and its subdirectories.
-Dir["#{File.dirname(__FILE__)}/support/**/*.rb"].each {|f| require f}
-
-RSpec.configure do |config|
-  # == Mock Framework
-  #
-  # If you prefer to use mocha, flexmock or RR, uncomment the appropriate line:
-  #
-  # config.mock_with :mocha
-  # config.mock_with :flexmock
-  # config.mock_with :rr
-  config.mock_with :rspec
-
-  config.fixture_path = "#{::Rails.root}/spec/fixtures"
-
-  # If you're not using ActiveRecord, or you'd prefer not to run each of your
-  # examples within a transaction, comment the following line or assign false
-  # instead of true.
-  config.use_transactional_fixtures = true
+require 'simplecov'
+SimpleCov.start do
+  add_filter '/spec/'
+  add_filter 'lib/spree_faq/engine'
+  add_group 'Controllers', 'app/controllers'
+  add_group 'Overrides', 'app/overrides'
+  add_group 'Models', 'app/models'
+  add_group 'Libraries', 'lib'
 end
 
+ENV['RAILS_ENV'] = 'test'
+
+require File.expand_path('../dummy/config/environment.rb', __FILE__)
+
+require 'rspec/rails'
+require 'shoulda-matchers'
+require 'ffaker'
+
+RSpec.configure do |config|
+  config.mock_with :rspec
+  config.use_transactional_fixtures = false
+  config.treat_symbols_as_metadata_keys_with_true_values = true
+end
+
+Dir[File.join(File.dirname(__FILE__), 'support/**/*.rb')].each { |f| require f }
